refactor app to prepare for further changes
