+++ pkg/libs/esrgan/esrgan_remove.go
