+++ pkg/state/statemachine/cli_state.go
