+++ pkg/utility/cli/cli_clear.go
